requirements.txt
crewai
langchain
anthropic
python-dotenv
PyGithub